+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- (Checkbox, Radio): Customizable theme colors
 - (form): Add CustomField component as a formik Field wrapper
 - (CheckboxField, InputField, RadioField, TextareaField): Forward `id` attribute
 - (CheckboxField): Forward `value` attribute (for checkbox arrays)